Added new opengl project settings, added some ifdefs

+++ Graphics/src/GraphicsInputLayout.cpp
@@ -48,7 +48,7 @@ namespace pawn {
 	}
 
 	void GraphicsInputLayout::RecalculateStride() {
-		size_t offset = 0;
+		uint32_t offset = 0;
 		m_Stride = 0;
 
 		for (auto& element : m_Elements) {
